riite: cast DBToAmp to float in the IBufferSender ctor
IBufferSender's constructor assigned the double returned by DBToAmp
straight into the float mThreshold, tripping C4244 on MSVC. The two
sibling sender classes in this same file already cast it explicitly
(ISender.h:129 and :238) - IBufferSender just missed it, and it only
surfaces now because a template member compiles when something
instantiates it. ColorSynth's 1024-sample scope sender is the first
instantiation in the house.

Behaviour is unchanged: the default -90 dB threshold is 3.1623e-5
either way, and mThreshold is only ever compared against a float
running sum.

+++ IPlug/ISender.h
@@ -442,7 +442,7 @@ public:
     if (minThresholdDb <= kNoThresholdDb)
       mThreshold = -1.0f;
     else
-      mThreshold = DBToAmp(minThresholdDb);
+      mThreshold = static_cast<float>(DBToAmp(minThresholdDb));
     
     SetBufferSize(bufferSize);
   }
